requirements.txt
telethon
python-dotenv
pydantic 